added camera controls
need to fix left and right movement to be based on current right.

Also need to clean up code everything is a mess and ugly, and slowly leading me to a state of insanity where I must find a specific line of code in amongst a jungle of words a numbers that cause my poor dyslexic brain pain and agony.

Consider adding roll to camera, although it's will be usless.